Implement signing without library.
This commit removes the reliance on Namshi/JOSE for JWT handling. I
spent some time trying to get the requests to work against the
boulder instance that I set up and it just was not working. After
spending some time looking at the Node application in the boulder
repo, I think that either Namshi/JOSE is building JWT requests
incorrectly, or the Let's Encrypt ACME server is a special snowflake
that needs a different JWT. I'm inclined to think that there is an
issue with Namshi/JOSE (or perhaps our implementation).

The biggest issue was that the header that is used to create the
protected field *requires* that you set an "alg" value. Namshi/JOSE
uses "alg" to determine how to sign the header/payload combo. The
problem is that this value is also added to the protected header.
Boulder's test Node application only includes the nonce in the
protected header. With this extra value, the requests were summarily
rejected.

As such, I started trying to build the request based on the test Node
application. I was able to steals bits from Namshi/JOSE (e.g., the
functions used for the actual signature), while ensuring that the
contents in the request correct.

In the end, the request objects are now readable by the ACME server.
We are still dependent upon Namshi/JOSE for the encoder/decoder;
however, this is easily replaceable if we decide to fully abandon
it.