Change "body" to "payload"
With the requests that are made, we are sending a JWT as the body of
the request; however, the object that we create need to be passed the
plaintext payloads that eventually get encrypted as JWTs. To clarify
that the payload and body are different things, change the "body"
language to "payload". "body" will be reintroduced to actually mean the
full body of the request.