Restructure Request args for flexibility
When working through the various requests that are needed to satisfy
the ACME protocol, it became clear that the current Request class is far
too rigid for the needs of the different requests. The class has been
restructured to allow for more flexibility in order to share a lot of the
functionality while allowing for the class to handle a wide array of the
requests being made.